create unlikes & fix likes double like bug
